+++ capstone-back-end-java/src/main/java/com/capston/project/back/end/repository/TokenRepository.java
@@ -6,9 +6,10 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
-	Optional<Token> findByToken(String token);
+public interface TokenRepository extends JpaRepository<Token, Long> {
+  Optional<Token> findByToken(String token);
 
-	@Query("SELECT t FROM Token t WHERE t.user.id = :userId AND t.expired = false AND t.revoked = false")
-	List<Token> findAllValidTokensByUser(Integer userId);
+  @Query(
+      "SELECT t FROM Token t WHERE t.user.id = :userId AND t.expired = false AND t.revoked = false")
+  List<Token> findAllValidTokensByUser(Long userId);
 }
